+++ apps/playground/src/app/components/indicators/indicators.module.ts
@@ -1,0 +1,25 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+
+const routes: Routes = [
+  {
+    path: 'status-indicator',
+    loadChildren: () =>
+      import('./status-indicator/status-indicator.module').then(
+        (m) => m.StatusIndicatorModule
+      ),
+  },
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class IndicatorsRoutingModule {}
+
+@NgModule({
+  imports: [IndicatorsRoutingModule],
+})
+export class IndicatorsModule {
+  public static routes = routes;
+}