+++ libs/components/lookup/src/lib/modules/lookup/lookup.component.ts
@@ -566,6 +566,7 @@ export class SkyLookupComponent
           useValue: contextProviderValue,
         },
       ],
+      size: 'large',
       wrapperClass: this.wrapperClass,
     });
   }