+++ libs/sdk/e2e-schematics/jest.config.ts
@@ -1,0 +1,25 @@
+/* eslint-disable */
+export default {
+  displayName: 'e2e-schematics',
+  preset: '../../../jest.preset.js',
+  globals: {
+    'ts-jest': {
+      tsconfig: '<rootDir>/tsconfig.spec.json',
+    },
+  },
+  transform: {
+    '^.+\\.[tj]s$': 'ts-jest',
+  },
+  moduleFileExtensions: ['ts', 'js', 'html'],
+  collectCoverage: true,
+  coverageDirectory: '../../../coverage/libs/sdk/e2e-schematics',
+  coverageThreshold: {
+    global: {
+      branches: 100,
+      functions: 100,
+      lines: 100,
+      statements: 100,
+    },
+  },
+  silent: true,
+};